requirements.txt
PyYAML>=6.0.1
dbt-duckdb>=1.6.0
